requirements.txt
requests>=2.31.0
dnspython>=2.4.2
ipwhois>=1.2.0
discord.py>=2.3.2
aiohttp>=3.9.1
argparse>=1.4.0
colorama>=0.4.6
tqdm>=4.66.1
python-nmap>=0.7.1
beautifulsoup4>=4.12.2 